sandbox/sandbox.py: Display failed commands in the detail string
We should only display commands in detail strings, not in the
message texts.

This also updates tests/integration/sandbox-bwrap.py to expect
the new message string which only contains the command exit status
and not the whole command itself, this does not alter the validity
of the text case which is checking that we can obtain the expected
return value.