tests/integration/pullbuildtrees.py: Fix the non-integration case.
This test has one test case which is marked as an integration test,
and the other is not an integration test, but was using the integration
cli. The integration cli does not work correctly if not run in
integration mode.

This was causing an error locally in conftest.py when trying to
create a tmpdir inside a nonexisting integration cache directory.