Add tox.ini from master branch
`tox.ini` has seen heavy churn lately, and backporting individual
commits is turning out to be a more involved process as the two branches
have diverged a bit. So, this is more or less copy-pasted from the
master branch, without a separate lint job.

Since switching to `pycodestyle` will require some code formatting
changes, that does not seem reasonable to apply to this branch. So,
let's stick with running the linters via `pytest`.