contrib/bst-here: Allow users to specify image variant
Now that the `buildstream/buildstream` image has 9 variants, let's make
it easier to choose the desired tag, using a command-line option.

This is otherwise possible by specifying the full image name `image:tag`
using the `-i` option. But, this will make it easier to specify just the
tag using `-j`.

The following two invocations of `bst-here` are now equivalent:

    bst-here -i buildstream/buildstream:dev
    bst-here -j dev