element.py: Force resolve intermediate dependency cache keys
When a cache key is discovered (e.g.: due to a workspace build completing),
then we have no guarantee that reverse dependency cache keys will be
updated in order.

As such, we must ensure that `Element._update_state()` serializes this
and ensures that cache keys for intermediate dependencies get resolved.

This fixes #919