cascache: limit 'infinite' cache to volume size
Make sure we don't exceed the size of the volume when calculating an
'unlimited' quota.

Note that at present, cache_size is an estimate that will be the same
as the cache_size or larger. In my own case, I have a volume dedicated
to the cache, so it was easy for the cache_quota to exceed this. This
prevented me from running BuildStream, although I had plenty of free
space.