runcli: rm unused 'color' and '**extra' params
The only customer for 'invoke' is the 'run' method on the same class.
Make it private and simplify it's interface.

Avoid disappointment by removing unused 'color' and 'extra' parameters
from 'invoke'.

These appear to have been here to mirror the interface of
click.testing.CliRunner.invoke(), which this class replaced.