Let subprocess decode stdout based on locale
Subprocesses can return decoded strings if we give them a
"universal_newlines=True" argument. We can therefore offload that to
them, and not explicitly decode output ourselves.

This also fixes multiple bugs where we would not be respecting the
locale used by the user, and in some cases force it to "ascii".