_gitsourcebase.py: Make mirror class used by derived plugins overridable
We want it to be possible for plugin authors writing git type plugins
derrived from _GitSourceBase to be able to make use of custom
_GitMirror classes.

Add a MirrorClass class variable to _GitSourceBase which can be
overriden by derriving classes. This allows plugin authors to specify
the use of custom/derrived _GitMirror class.