Rename cache-buildtrees option value 'failure' to 'auto'
We anticipate other cases than build failures where buildtree caching
will be required.  E.g., incremental workspace build with remote
execution. Or running tests in a buildtree in parallel with the build of
reverse dependencies.

This renames the option value 'failure' to the more generic 'auto' to
cover these other cases as well.