_yaml.py: Rip out ChainMap(), node_chain_copy(), node_list_copy()
This class and these two functions exist as they were intended to
bring efficiency. Benchmarking this patch against the debian-stack.bst
element in the debian-like project [0] showed that although this
took 15M more RAM (peak usage), there was a ~20s gain in the time taken
to 'show' the stack. Thus the class and functions have been removed.

This also has the advantage of removing a lot of duplicate and unnecessary
code.

[0] https://gitlab.com/jennis/debian-stretch-bst