_loader/loader.py: Specify junction name in missing file errors where appropriate
When a file is missing in a subproject, it is not particularly meaningful
to specify the filesystem path to the elements directory of the subproject,
as this temporary staging directory belongs to BuildStream and not the user.

Instead, when a file is missing in a subproject, specifying the junction name
is more useful.

This fixes an aspect of #947