Element.__init_defaults: init if no plugin_conf
Still initialize `Element.__defaults` as usual, even if a `plugin_conf`
is not specified.

As part of the work on !1101 to introduce `ArtifactElement`s, a special
case for `__init_defaults` was made, such that none of the usual
defaults would be loaded. This seems to be because it has no config and
none of the defaults are required by ArtifactElement, although they
don't seem to adversely affect it.

Instead, treat a 'None' `plugin_conf` in the same way as the existing
case of a missing file. This avoids the appearance of having
plugin-specific behaviour in the base-class, and is perhaps less
puzzling to new contributors.