cascache: atomically save size via tmpdir instead
Instead of doing the atomic 'mv into place' using a temporary file in
the 'cas' directory, use the sibling 'tmp' directory instead.

This will avoid the situation where CacheSizeJob encounters a temporary
file in the cas directory. This can happen because CacheSizeJob does not
always run with an exclusive lock on the cas directory, and BuildQueue
updates the cache size whenever a BuildJob completes.