_artifact.py: Do not create empty buildtree directories
Creating an empty buildtree directory when buildtree caching is disabled
means cached_buildtree() can return True even if no buildtree is
available, which breaks logic in the frontend and in Stream.

This adds the buildtree_exists() method to indicate whether an artifact
was created with a buildtree, independent of whether the buildtree is
available in the local cache.