_project: don't _find_project_dir if a junction
If we're creating a project for a junction, then don't go looking for
the project directory - the path has already been given to us.

This means that we can now detect when junctions are missing a
'project.conf', and importantly we don't accidentally start resolving
elements in the enclosing project.

Add tests to cover workspaced, local, and git repo cases. Note that
this is also the first test coverage for the INVALID_JUNCTION path.

In later work we might extract the _find_project_dir magic out of the
Project class, so that there are no surprises when instantiating it.