loader: clearer error if no junction project.conf
I found the original message confusing - it would talk about wanting a
project at '.', without a frame of reference. After re-reading the code
and junction docs it made more sense, I think this version would save
folks the trouble.

Removing all mention of it, if it's not specified. If it is specified
then explain it a bit more, i.e.

Before:

    Could not find the project.conf file for junction element at
    hello-junction.bst [line 1 column 0]. Expecting a project at path
    'project'

After:

    Could not find the project.conf file in project referred to by
    junction element 'hello-junction.bst'. Was expecting it at path
    'project' in the junction's source.

If no 'path' specified in the junction:

    Could not find the project.conf file in project referred to by
    junction element 'hello-junction.bst'.