app.py: str(e) instead of "{}".format(e)
We're not implementing __format__ anywhere, so there's no reason to
prefer format() over the more conventional conversion with str().

In one case no conversion is needed at all, so remove it.

In another case the conversion is better done in the called function
(_notify), not where the function is called.

For the _message case, there are quite a few places that might need
fixing up, so that is left for later work.

Note that SubprocessError is a vanilla subclass of Exception:
https://github.com/python/cpython/blob/bafa8487f77fa076de3a06755399daf81cb75598/Lib/subprocess.py#L96