tests: Move conftest.py into tests/ subdirectory
Pytest implicitly loads this file from the CWD where it runs,
which is usually the project toplevel directory, but also supports
loading it from the `tests/` subdirectory where tests are run from.

Placing it into the `tests/` subdirectory is not perfectly explicit,
but is at least a hint to the unsuspecting developer that this file
is related to tests.