tests: ensure provenance for _get_loader errors
Make the tests that currently cover _get_loader ensure that we are
getting the expected provenance.

Note that for some tests, we must use yaml_file_get_provenance, as the
generated yaml is not stable across versions of ruamel.

In later work we may replace all instances of provenance string tests
with yaml_file_get_provenance, as it will be more robust to future
changes.