setup.py: Allow optionality of main entry point at install time.
By default, we now install as `bst2` so as to not overlap with
`bst` which is BuildStream 1.

However, as it does not pose any problem to install as `bst`, we
allow it. The expectation is that once 2.0 is released, distros will
probably install `bst1` and `bst2` separately, but alias the `bst`
command to the latest version (e.g., like the debian 'alternatives' concept).