spawn: split jobs into parent and child
Make it a bit clearer what happens in which process by splitting out a
'ChildJob', which encapsulates the work that will happen in the child
process.

This also makes it possible to control what will be transferred to the
child process, which is very useful for adding support for the 'spawn'
method of creating child processes as opposed to the 'fork' method.