_scheduler/jobs: split jobs into parent and child
Make it clearer what happens in which process by splitting out a
'ChildJob', which encapsulates the work that happens in the child
process.

This also makes it possible to control what is transferred to the
child process. This is very useful for adding support for the 'spawn'
method of creating child processes as opposed to the 'fork' method.