_project.py: Do not find default targets in .bst directory
In cases where element path is set to `.` and the project has some
junctions, we have to be careful when trying to find the default
targets. We can't simply find all files ending with `.bst`. But, we
first need to ensure that we do not recurse down the `.bst` directory
which is used to store BuildStream internals.

Fixes #992.