Add support for defining target for junction elements
Add support for specifying target of a junction element to be a
sub-sub-project as a configuration option. When such a target is
defined, the junction element is not allowed to have any sources.

This was discussed in the following mailing list thread:
https://mail.gnome.org/archives/buildstream-list/2019-April/msg00025.html

Since this adds new configuration, also bump `BST_FORMAT_VERSION`.