_artifactcache/casserver.py: Add message handler
This needs to be added along with the status messages added
to the artifact cache, and this detail diverges from master.

This is because we have not bothered to backport !1071 which
refactors the CASCache to be a delegate object of the ArtifactCache
istead of a derived class - backporting !1071 would allow us to
remove these message handlers because the CAS server and test
fixture only use the CASCache object directly, not the business
logic in the ArtifactCache.