HACK: unsuccessful multi-venv experiment
This is an experiment to test the feasability of supporting a venv per
plugin, all running under the same interpreter.

It shows:

  * Installing two separate venvs with different dependency versions
  * Running an interpretor which loads a plugin in both separate venvs
    (it could even be the same plugin, and need not be a "BuildStream"
    plugin, but some python module loaded on demand)
  * Prove that we infact have separation (perhaps by having the plugin
    just print the versions of it's dependencies).

The approach taken in this experiment is to push all modules required by the
plugin into a PluginBase space.

There are major problems with this approach, see the README.md.