doc: Fix the build with Sphinx 2.0
Sphinx 1.7, released in February 2018 moved the sphinx.apidoc module to
sphinx.ext.apidoc, with an alias and a deprecation warning in place so
users know to port their code.

The compatibility alias was removed in Sphinx 2.0, so we need to move to
the new module name.

Fortunately, since the new module name is more than a year old, this
shouldn't break anything for anybody.