Revert "_stream.py, _project.py: Manage GC during pipeline load"
This change was initialy introduced as it provided a minor gain in the
load time of the elements, pre-scheduler.

For the base-files subset of the debian-stack, this gain was ~2 seconds.
However, with it, came a fairly significant cost to peak memory usage (~100M).

Additionally, the benchmarks have highlighted that this patch has introduced
some cost during build time, where the build of the base-files subset of the
debian-like project has increased by ~90 seconds and peak memory by ~100M

This reverts commit 746aa7a63c4d06b1caa012a35c970b43e0166faa and
commit d55039c33685d267fd8834ecc5d16030c1385325.