plugin.py: Start plugin ID counter at ID 1
This was always intended, but was not well commented. The reason
we start plugin ID counters at 1 is that we prefer relying on
a falsy value to determine whether an ID holding variable has
been set or not.

This patch also adds a more informative assertion in Plugin._lookup()

This by itself essentially fixes #1012