testing._sourcetests: Don't special case 'local' in parameter list
The 'local' kind is hard coded in the parameter list of on test in
testing._sourcetests as so will always run regardless of what plugins
have been registered.

Remove this special casing by duplicating the test in the local source
specific test.

Ideally, the local source should have a Repo implementation and be
registered with the templated tests.