element.py: Delegate cache key calculation to CacheKeyController objects
This commit adds a CacheKeyController class (with subclasses
StrictCacheKeyController and NonStrictCacheKeyController), which are in
charge of storing and generating cache keys.

The Element now stores a __strict_cache_key_ctrl and a __cache_key_ctrl
(when buildstream runs in strict mode, these are the exact same object),
and calls out to these to calculate and retrieve cache keys.