source.py: ensure previous track refs are updated
Updates the refs in the job process (but doesn't write), to ensure
following sources can see consistency of previous sourcse has been
updated. `_save_ref` is renamed `_set_ref` with writing to file now
optional.

This also changes the previous_source_access test to use a remote, so
that it actually tests this cornercase.

Fixes #1010