jobs: refactor, use new set_message_unique_id
Ease the burden on subclasses of Job slightly, by providing a new
set_message_unique_id() method. It ensures that created Message
instances will use that id.

This removes the need to override the message() method, so it is no
longer in the 'abstract method' section.

Enable callers to Job's message() method to override the 'unique_id'.