_loader/loader.py: Set junction to be required in order to resolve cache key
Recent code changes has avoided resolving cache keys in some conditions,
which causes loading to break when attempting to stage junctions in non-strict
mode.

Instead of calling Element._update_state() here to resolve the
cache key, setting it to be required causes it to automatically
have it's cache key resolved, solving the problem.

This change fixes #1018