element.py: Update state on reverse dependencies when strict key is resolved.
When calculating the strict key, we do not know for sure yet that the
element is required or whether it will ever have a cache key, as
we could be tracking elements which will result in resolving an
element which is a build-only dependency of the target.

This ensures that we process all elements which need to be processed,
which is important for cases where we are building in non-strict mode
and tracking is also enabled.

This fixes issue #1014