New test stage to use the buildbox sandbox
This test stage should only be needed while we transistion to the
BuildBox sand box. Once the other test stages switch to use buildbox
from bubble rap then this test stage will be removed.

These test do not include the integration test as thery are not working
correctly in CI.