_yaml: Cythonize and internalize Node
Node used to be a NamedTuple that we used to access by index
for speed reasons. Moving to an extension class allows us to
access attributes by name, making the code easier to read and
less error prone. Moreover, we do gain speed and memory by
doing this move.

Also fix a few places where we would not have an entire `Node`
object but were instead just returning a tuple, missing some entries.