Introduce Cython to the project and document
Cython requires a plugin to allow coverage of cython files, which
was updated in coveragerc. It also means we need to build the
dependencies and install cython for coverage.

Cython requires access to both source and compiled files when
running coverage. We therefore need to install project in develop
mode.

Updated documentation to explain how to run tests without tox but
with coverage