Rename (spawn, fork) -> 'start process'
Avoid confusion by not referring to starting another process as
'spawning'. Note that 'spawn' is a process creation method, which is an
alternative to forking.

Say 'create child process' instead of 'fork' where it doesn't harm
understanding. Although we currently only use the 'fork' method for
creating subprocesses, there are reasons for us to support 'spawn' in
the future.

More information on forking and spawning:
https://docs.python.org/3/library/multiprocessing.html#contexts-and-start-methods