_fuse: converge args for link() and symlink()
Get consistency between the Operations base class and the
SafeHardlinkOps subclass.

Use the 'target, source' convention as it appears in Operations.

This avoids the confusion of 'target' changing meaning in the
'name, target' convention introduced by SafeHardlinkOps. The naming is
also consistent with 'man ln', and more consistent with Python's
equivalent functions.

Unfortunately FUSE.link() and Operations.link() swap the order of
'target' and 'source', but the meaning is the same.