_options: match methods to base class
Update OptionEnum and OptionFlags to match the Option base-class.

Introduce 'load_special' to allow explicit usage of extended
functionality, which wouldn't necessarily be supported by different
subclasses.

By making the signatures of overridden methods deliberately match, we
can use PyLint to ensure the signatures don't accidentally differ.

Also replace `super(cls, self)` with the equivalent and less error-prone
`super()`, wherever we've touched it. In separate work we may want to
replace this across the project.

We must introduce __init__() functions for OptionEnum and OptionFlags to
get a clean pass from PyLint. It seems that it's unable to see that
load_special() is also transitively called from __init__().