scheduler.py: Introduce a notification queue for stream interaction
For the 'backend' of BuildStream to run in a subprocess certain
interactions between the 'frontend' need to be augmented over a
queue between the processes, as well as adaptions for the differences
in state between them. This commit implements the core of these
interactions as notification objects over a deque with notifying
callbacks on either side, which should be switched to a
multiprocessing queue once there's process separation at which
point 'both ends' can execute an event loop.

This also removes the checks for the ticker & interrupt callbacks
to be optional, as currently they're always implemented.