Allowing the tar plugin to use a local path
_downloadablefilesource.py: Added ability to download from either url or path

Added changes to the code so that now a downloadable file can be sourced
from either a given url or a local path. If both are given, error occurs.

tar.py: Made changes to accept the path sourcing

When a path is given, the tar can be found from the local directory
rather than the mirror directory. Made changes to make the code less
specific to urls so that now it will accept both url and path.