_yaml: Add a 'get_node' on Mapping, when return type can be of multiple types
Sometimes, we might want to have a list or a string, in which case, we
can now use 'get_node', which can return multiple types.

This method doesn't contain a 'default' value, as building a 'Node' from
it would be hard. We therefore just have a 'allow_none' flag that allows
returning 'None' when the value is not found.

- includes: use the new 'get_node' instead of trying multiple versions