platform: re-scope set_resource_limits
Rename 'set_resource_limits' to 'maximize_open_file_limit', as this
seems to more accurately reflect it's function.

Remove unused flexibility from the implementation, to make it easier to
understand.

Simplify the Mac implementation, and add some explanation for the
OPEN_MAX magic number. In later work we should remove the magic number.

Import 'resource' late, which is not available on Windows.